Refactor into impl_
